wallet/requirements.txt
bit
eth-account
python-dotenv
web3